extracts cube from url for logic layer
