+++ tesseract-server/src/handlers/logic_layer/mod.rs
@@ -1,6 +1,6 @@
-mod detection;
+mod aggregate;
 pub mod shared;
 
-pub use self::detection::logic_layer_handler;
-pub use self::detection::logic_layer_default_handler;
+pub use self::aggregate::logic_layer_handler;
+pub use self::aggregate::logic_layer_default_handler;
 pub use self::shared::{Time, TimePrecision, TimeValue, LogicLayerQueryOpt, finish_aggregation};
